requirements.txt
appdirs==1.4.1
certifi==2018.1.18
chardet==3.0.4
Django==1.9
django-cors-headers==2.0.2
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
gunicorn==19.6.0
idna==2.6
packaging==16.8
pkg-resources==0.0.0
psycopg2==2.7.4
PyJWT==1.4.2
pyparsing==2.1.10
pytz==2018.3
requests==2.18.4
six==1.10.0
urllib3==1.22
